OP-777: Fixed UploadClaimsTestCase unit tests
