+++ tools/tests/__init__.py
@@ -1,0 +1,1 @@
+from tools.tests.test_services import *
